Create README
## Flexxia Solution
Hi, thanks for reviewing my code!

Just a quick note: The charts resize correctly when dragging the window pane, but not always when using the maximize button to return to a non-maximized size. I tested without the charts, and all of my elements are resizing correctly in both cases, so it may be  an issue with the library. Another bug I encountered is if you quickly resize the window pane while the chart is rendering, it will render more than one chart at different sizes. 

Encountered both issues using Firefox 54.0.1